Split out favicon component
