+++ frontend/src/pages/sites/siteForm/SiteForm.tsx
@@ -1,7 +1,7 @@
 import { yupResolver } from "@hookform/resolvers/yup";
 import cx from "classnames";
 import { capitalize } from "lodash-es";
-import { type ChangeEvent, type FC, useState } from "react";
+import { type FC, useState } from "react";
 import { Button, Form } from "react-bootstrap";
 import { Controller, useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
@@ -9,18 +9,15 @@ import Select from "react-select";
 import {
   type SiteCreateInput,
   type SiteQuery,
-  useLazyFetchSiteFavicons,
   useSiteCategories,
   ValidSiteTypeEnum,
 } from "src/graphql";
 import * as yup from "yup";
+import SiteFavicon from "./SiteFavicon";
 
 type Site = NonNullable<SiteQuery["findSite"]>;
 
 const validSites = Object.keys(ValidSiteTypeEnum);
-
-const parseMimeType = (dataURL: string) =>
-  dataURL.match(/^data:([^;,]+)/)?.[1] ?? "unknown";
 
 const schema = yup.object({
   name: yup.string().required("Name is required"),
@@ -57,44 +54,6 @@ const SiteForm: FC<SiteProps> = ({ site, callback }) => {
 
   // null: favicon unchanged, "": cleared, otherwise a base64 data URL to store.
   const [favicon, setFavicon] = useState<string | null>(null);
-  const [candidates, setCandidates] = useState<
-    { url: string; image: string }[]
-  >([]);
-  const [faviconError, setFaviconError] = useState<string>();
-  // Pixel dimensions per candidate, read from each image once it loads.
-  const [dimensions, setDimensions] = useState<Record<string, string>>({});
-  const [fetchFavicons, { loading: fetchingFavicons }] =
-    useLazyFetchSiteFavicons();
-
-  const currentIcon = favicon === null ? (site?.icon ?? "") : favicon;
-
-  const handleFetchFavicons = () => {
-    setFaviconError(undefined);
-    setCandidates([]);
-    setDimensions({});
-    fetchFavicons({ variables: { url: getValues("url") ?? "" } })
-      .then((res) => {
-        if (res.error) {
-          setFaviconError(res.error.message);
-          return;
-        }
-        const found = res.data?.fetchSiteFavicons ?? [];
-        if (found.length === 0) setFaviconError("No favicons found");
-        setCandidates(found.map(({ url, image }) => ({ url, image })));
-      })
-      .catch((e: unknown) => {
-        setFaviconError(e instanceof Error ? e.message : "Failed to fetch");
-      });
-  };
-
-  const onFileChange = (event: ChangeEvent<HTMLInputElement>) => {
-    if (event.target.validity.valid && event.target.files?.[0]) {
-      const reader = new FileReader();
-      reader.onload = (e) =>
-        e.target?.result && setFavicon(e.target.result as string);
-      reader.readAsDataURL(event.target.files[0]);
-    }
-  };
 
   const categories = (
     categoryData?.querySiteCategories.site_categories ?? []
@@ -246,89 +205,12 @@ const SiteForm: FC<SiteProps> = ({ site, callback }) => {
         </Form.Text>
       </Form.Group>
 
-      <Form.Group className="mb-3">
-        <Form.Label>Favicon</Form.Label>
-        <div className="d-flex align-items-center gap-2 mb-2">
-          {currentIcon ? (
-            <img src={currentIcon} alt="" width={24} height={24} />
-          ) : (
-            <span className="text-muted">No favicon</span>
-          )}
-          <Button
-            type="button"
-            size="sm"
-            onClick={handleFetchFavicons}
-            disabled={fetchingFavicons}
-          >
-            {fetchingFavicons ? "Fetching…" : "Fetch favicons"}
-          </Button>
-          {currentIcon && (
-            <Button
-              type="button"
-              size="sm"
-              variant="outline-danger"
-              onClick={() => {
-                setFavicon("");
-                setCandidates([]);
-              }}
-            >
-              Remove
-            </Button>
-          )}
-          <Form.Control
-            type="file"
-            size="sm"
-            className="w-auto ms-auto"
-            onChange={onFileChange}
-            accept=".ico,.png,.jpg,.jpeg,.webp,.svg,.gif"
-          />
-        </div>
-
-        {faviconError && (
-          <div className="text-danger small mb-2">{faviconError}</div>
-        )}
-
-        {candidates.length > 0 && (
-          <div className="d-flex flex-wrap gap-2 mb-2">
-            {candidates.map((candidate) => (
-              <button
-                type="button"
-                key={candidate.url}
-                className={cx(
-                  "btn btn-outline-secondary d-flex flex-column align-items-center p-2",
-                  { active: favicon === candidate.image },
-                )}
-                title={candidate.url}
-                onClick={() => setFavicon(candidate.image)}
-              >
-                <img
-                  src={candidate.image}
-                  alt=""
-                  width={32}
-                  height={32}
-                  onLoad={(e) => {
-                    const { naturalWidth, naturalHeight } = e.currentTarget;
-                    setDimensions((d) => ({
-                      ...d,
-                      [candidate.url]: `${naturalWidth}×${naturalHeight}`,
-                    }));
-                  }}
-                />
-                <span className="small text-muted mt-1">
-                  {parseMimeType(candidate.image)}
-                  {dimensions[candidate.url]
-                    ? ` · ${dimensions[candidate.url]}`
-                    : ""}
-                </span>
-              </button>
-            ))}
-          </div>
-        )}
-
-        <Form.Text>
-          Fetch favicons from the site URL, or upload a custom icon.
-        </Form.Text>
-      </Form.Group>
+      <SiteFavicon
+        currentIcon={site?.icon}
+        getUrl={() => getValues("url") ?? ""}
+        value={favicon}
+        onChange={setFavicon}
+      />
 
       <Form.Group className="d-flex mb-3">
         <Button type="submit" className="col-2">
